+++ wandb/util.py
@@ -157,6 +157,31 @@ class LazyLoader(types.ModuleType):
         module = self._load()
         return dir(module)
 
+class PreInitObject(object):
+    def __init__(self, name):
+        self._name = name
+
+    def __getitem__(self, key):
+        raise wandb.Error(
+            'You must call wandb.init() before {}["{}"]'.format(self._name, key))
+
+    def __setitem__(self, key, value):
+        raise wandb.Error(
+            'You must call wandb.init() before {}["{}"]'.format(self._name, key))
+
+    def __setattr__(self, key, value):
+        if not key.startswith("_"):
+            raise wandb.Error(
+                'You must call wandb.init() before {}.{}'.format(self._name, key))
+        else:
+            return object.__setattr__(self, key, value)
+
+    def __getattr__(self, key):
+        if not key.startswith("_"):
+            raise wandb.Error(
+                'You must call wandb.init() before {}.{}'.format(self._name, key))
+        else:
+            return object.__getattr__(self, key)
 
 np = get_module('numpy')
 
@@ -740,9 +765,6 @@ def guess_data_type(shape):
     # (samples,) or (samples,logits)
     if len(shape) in (1, 2):
         return 'label'
-    # (samples, height, width) = grayscale image
-    if len(shape) == 3:
-        return 'image'
     if len(shape) == 4:
         if shape[-1] in (1, 3, 4):
             # (samples, height, width, Y \ RGB \ RGBA)